+++ src/test/resources/io/jenkins/plugins/analysis/warnings/recorder/make-gcc/main.c
@@ -2,5 +2,6 @@
 
 int main() {
     printf("Michi and Raphi say hello jenkins\n");
+    printf(42);
     return 0;
 }
